# oremedia/docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oremedia application API (tRPC procedure catalogue)",
    "version": "1"
  },
  "paths": {
...
    "/trpc/brand.versions.publish": {
      "post": {
        "operationId": "brand.versions.publish",
        "x-trpc-type": "mutation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brandId": {
                    "type": "string"
                  },
                  "versionId": {
                    "type": "string"
                  },
                  "expectedVersion": {
                    "type": "number"
                  }
                },
                "required": ["brandId", "versionId", "expectedVersion"]
              }
            }
          }
        }
      }
    },
...
    "/trpc/creative.comments.resolve": {
      "post": {
        "operationId": "creative.comments.resolve",
        "x-trpc-type": "mutation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentId": {
                    "type": "string"
                  },
                  "commentId": {
                    "type": "string"
                  },
                  "expectedVersion": {
                    "type": "number"
                  }
                },
                "required": ["documentId", "commentId", "expectedVersion"]
              }
            }
          }
        }
      }
    },
...
    "/trpc/creative.renders.request": {
      "post": {
        "operationId": "creative.renders.request",
        "x-trpc-type": "mutation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "documentId": {
                    "type": "string"
                  },
                  "revisionId": {
                    "type": "string"
                  },
                  "formatKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["documentId", "revisionId", "formatKeys"]
              }
            }
          }
        }
      }
    },
...
    "/trpc/operations.deletion.request": {
      "post": {
        "operationId": "operations.deletion.request",
        "x-trpc-type": "mutation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "subjectType": {
                    "type": "string",
                    "enum": ["user", "asset", "brand", "tenant", "channel_connection", "customer_voice"]
                  },
                  "subjectId": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": ["subjectType", "subjectId", "reason"]
              }
            }
          }
        }
      }
    },
...
    "/trpc/skills.versions.publish": {
      "post": {
        "operationId": "skills.versions.publish",
        "x-trpc-type": "mutation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "skillVersionId": {
                    "type": "string"
                  },
                  "expectedVersion": {
                    "type": "number"
                  },
                  "rolloutPercent": {
                    "type": "number"
                  }
                },
                "required": ["skillVersionId", "expectedVersion"]
              }
            }
          }
        }
      }
    }
  },
...
}